Adaptation to the current state of the unification algorithm (most probably a consequence of the refolding of recursive constants).
A problem of the form

  Require Import List.
  Import ListNotations.
  Lemma foo l1 l2 : (forall t x y z : list nat, (t ++ x ++ y) ++ z = y ++ x) ->
    [1] ++ [0] ++ l1 ++ l2 = l1 ++ [0].
  intros. refine (H [_] [_] _ _).

is not working any more and most probably, it is because after some
step of unfolding the definition of ++ on t and x which are
respectively assigned to the singleton lists [1] and [0], one arrives
at the following problem

  [] ++ l1 ++ l2 = ([] ++ ?y) ++ ?z

which reduces to

  [] = [] ++ ?1 and l1 ++ l2 = ?z

which is solvable by taking ?y = [] and ?z = l1++l2.

Before, the leftmost ++ appeared as a Fix in the same problem, so it
eagerly continued reducing to l1 ++ l2 = ?y ++ ?z.

Note that in 8.4, the following variant did not work either

  Lemma foo l1 l2 : (forall x y z : list nat, (x ++ y) ++ z = y ++ x ++ z) ->
    [0] ++  l1 ++  l2 = l1 ++ [0] ++ l2.
  intros. refine (H [_] _ _).

So, the problem is probably more on the ability to backtrack from a
wrong unfolding in problems of the form "cst args = cst args2" than on
refolding or not recursive definitions, strictly speaking.

git-svn-id: svn://scm.gforge.inria.fr/svnroot/coq-contribs/trunk@2321 0cf17b13-060f-0410-b1b1-c666bec9822a